feat(row): convert to typescript
